requirements.txt
pandas
scikit-learn
joblib
streamlit
google-generativeai

